# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Polling Booth Officer Allocation & Notification API",
    "version": "1.0.0",
    "description": "REST API of the Smart Polling Booth Officer Allocation and Notification System.\n\n**Authentication** - single-login system (Main Admin).\n1. `POST /api/auth/login` with the credentials from `backend/.env` (`ADMIN_USERNAME` / `ADMIN_PASSWORD`).\n2. Copy `token` from the response.\n3. Click **Authorize** and paste it (Swagger adds the `Bearer ` prefix automatically).\n4. Mandal-scoped users only ever receive their own Mandal's data; the scope is forced server-side.\n\n**E-mail** - mail is sent through the Gmail API (OAuth2). Connect the mailbox once by opening `/auth/google` in a browser; the refresh token is stored in MongoDB, so sending also works after a server restart.\n\n**Raw JSON** of this specification is served from `/api-docs.json` (importable into Postman / Insomnia).",
    "contact": {
      "name": "Polling Officer Allocation System"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local backend - the port must match PORT in backend/.env"
    },
    {
      "url": "http://localhost:5002",
      "description": "Fallback port used when PORT is not defined"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/officers": {
      "get": {
        "tags": [
          "Officers"
        ],
        "summary": "List officers (paginated)",
        "description": "Mandal-scoped users automatically only receive officers of their own Mandal.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Search"
          },
          {
            "$ref": "#/components/parameters/MandalFilter"
          },
          {
            "name": "ward",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "designation",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated officer list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Officer"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Officers"
        ],
        "summary": "Create an officer",
        "description": "Roles: SUPER_ADMIN, ALLOCATION_OFFICER. `officerId` must be unique.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficerInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Officer created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficerResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/api/booths": {
      "get": {
        "tags": [
          "Booths"
        ],
        "summary": "List booths (paginated)",
        "parameters": [
          {
            "$ref": "#/components/parameters/Search"
          },
          {
            "$ref": "#/components/parameters/MandalFilter"
          },
          {
            "name": "ward",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated booth list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Booth"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Booths"
        ],
        "summary": "Create a booth",
        "description": "Roles: SUPER_ADMIN, ALLOCATION_OFFICER. `boothId` must be unique.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoothInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booth created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoothResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/api/allocation": {
      "get": {
        "tags": [
          "Allocation"
        ],
        "summary": "List allocations (paginated)",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALLOCATED",
                "CANCELLED",
                "REALLOCATED"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/MandalFilter"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated allocations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Allocation"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/dashboard/stats": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Dashboard statistics",
        "description": "Counts of officers, booths, allocations and notifications plus the required / available slots.",
        "responses": {
          "200": {
            "description": "Statistics object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/upload/officers": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Import officers from an Excel file",
        "description": "Super Admin only. Use `mode=preview` to validate the file without saving anything, `mode=commit` to import.",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "preview",
                "commit"
              ],
              "default": "commit"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Excel workbook (.xlsx / .xls), max 5 MB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import result (preview or committed rows)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or row validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/upload/booths": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Import booths from an Excel file",
        "description": "Super Admin only. Use `mode=preview` to validate the file without saving anything, `mode=commit` to import.",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "preview",
                "commit"
              ],
              "default": "commit"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Excel workbook (.xlsx / .xls), max 5 MB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import result (preview or committed rows)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or row validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Notification history (paginated)",
        "description": "SMS delivery status for every message sent to an officer. Mandal-scoped users only see their own Mandal.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "SENT",
                "FAILED",
                "DEMO_SENT"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/MandalFilter"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Notification"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/mail/send": {
      "post": {
        "tags": [
          "Mail"
        ],
        "summary": "Send an e-mail through Gmail",
        "description": "Roles: SUPER_ADMIN, ALLOCATION_OFFICER. `text` and/or `html` is required; when both are supplied the message is sent as multipart/alternative. Recipients may be a comma separated string or an array. Answers 503 while no Gmail account is connected (open /auth/google once).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "E-mail accepted by Gmail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailSendResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "502": {
            "description": "Gmail rejected the message (MAIL_SEND_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Gmail not configured or not connected (MAIL_NOT_CONFIGURED | MAIL_NOT_CONNECTED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/google": {
      "get": {
        "tags": [
          "Google Auth"
        ],
        "summary": "Start the Gmail OAuth2 consent flow",
        "description": "Redirects (302) the browser to the Google consent screen with `access_type=offline` and `prompt=consent`, so a refresh token is always issued. Open this URL once in a browser - it is not a JSON endpoint. The redirect URI must match GOOGLE_REDIRECT_URI in backend/.env.",
        "security": [],
        "responses": {
          "302": {
            "description": "Redirect to accounts.google.com",
            "headers": {
              "Location": {
                "description": "Google consent URL",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "GOOGLE_CLIENT_ID / GOOGLE_CLIENT_SECRET / GOOGLE_REDIRECT_URI missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "tags": [
          "Google Auth"
        ],
        "summary": "OAuth2 callback - stores the refresh token",
        "description": "Google redirects here after consent. The code is exchanged for tokens, the tokens are persisted in MongoDB (collection `googletokens`) and an HTML result page is returned. After a successful call, `POST /api/mail/send` works - including after a server restart.",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code supplied by Google"
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Set by Google when the user denied the request"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML result page (Gmail connected)",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Authorization code missing or consent denied (HTML page)",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Token exchange failed (HTML page)",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned by POST /api/auth/login"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid or expired JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request validation failed (express-validator)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorEnvelope"
            }
          }
        }
      },
...
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "1-based page number",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Records per page (max 500)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 50
        }
      },
      "MandalFilter": {
        "name": "mandal",
        "in": "query",
        "description": "Case-insensitive exact Mandal name (ignored for Mandal-scoped users, whose scope is forced server-side)",
        "schema": {
          "type": "string"
        }
      },
      "Search": {
        "name": "search",
        "in": "query",
        "description": "Free text search (ID, name, mobile number, locality)",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "description": "Standard success wrapper used by every endpoint",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 50
          },
          "total": {
            "type": "integer",
            "example": 412
          },
          "pages": {
            "type": "integer",
            "example": 9
          }
        }
      },
...
      "Officer": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "officerId": {
            "type": "string",
            "example": "OFF001"
          },
          "officerName": {
            "type": "string",
            "example": "Ramesh Kumar"
          },
          "name": {
            "type": "string",
            "description": "Alias of officerName"
          },
          "designation": {
            "type": "string",
            "example": "TA"
          },
          "mobileNumber": {
            "type": "string",
            "example": "9876543210"
          },
          "email": {
            "type": "string",
            "description": "Used by the e-mail endpoints - may be empty",
            "example": "ramesh@example.com"
          },
          "houseNo": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          },
          "mandalId": {
            "type": "string"
          },
          "mandal": {
            "type": "string",
            "example": "Kukatpally"
          },
          "district": {
            "type": "string"
          },
          "pinCode": {
            "type": "string",
            "example": "500072"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "OfficerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Officer"
          }
        }
      },
...
    }
  }
}